feat(runtime/hash): use `size_t` instead of `unsigned`

+++ src/runtime/hash.h
@@ -12,7 +12,7 @@ namespace lean {
 
 void mix(unsigned & a, unsigned & b, unsigned & c);
 
-unsigned hash_str(unsigned len, char const * str, unsigned init_value);
+unsigned hash_str(size_t len, char const * str, unsigned init_value);
 
 inline unsigned hash(unsigned h1, unsigned h2) {
     h2 -= h1; h2 ^= (h1 << 8);
